Move get_histograms_from_file to hist_io
